refactor: Phase 8 continued — extract useContextCutoff, useSessionPersistence, fold impersonation into useChatGeneration
- useContextCutoff: cutoff calc, cache, debounce, hide/unhide messages, context sheet, saveCurrentMessages
- useSessionPersistence: asyncSave, visibility save, image auto-hide, activeChar watch, beforeUnmount save
- startImpersonation moved into useChatGeneration
- ChatView: 2994 -> 2569 lines (-14.2%)
- Fix TDZ: isOpeningChat declared before useContextCutoff, asyncSaveCurrentSessionState passed as getter
